Fix all patch branch error on CentOS 7

+++ src/windtoolkit.cpp
@@ -205,7 +205,7 @@ wxString WindToolkitDialog::GetYear()
 wxArrayString WindToolkitDialog::GetHubHeights()
 {
 	wxArrayString hh;
-	for (int i = 0; i < lstHubheights->GetCount(); i++)
+	for (size_t i = 0; i < lstHubheights->GetCount(); i++)
 		if (lstHubheights->IsSelected(i))
 			hh.Add(lstHubheights->GetString(i));
 	return hh;
